% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunFor.R
\name{eigsmooth}
\alias{eigsmooth}
\title{Smoothing Function}
\usage{
eigsmooth(Y, pve = 0.9, npc = NULL)
}
\arguments{
\item{Y}{Y matrix}

\item{pve}{Proportion of variance explained}

\item{npc}{Number of PCs}
}
\description{
This function is a helper function called by make_split()
}
\examples{
eigsmooth(Y)
}
