% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunFor.R
\name{make_split}
\alias{make_split}
\title{Make a tree split}
\usage{
make_split(
  data,
  node,
  Xs,
  Ys,
  MVT,
  ancestor,
  depth,
  pred_list,
  npc,
  m_split,
  smooth
)
}
\arguments{
\item{data}{All the data}

\item{node}{Node to split}

\item{Xs}{Names of X variables}

\item{Ys}{Names of Y variables}

\item{MVT}{A matrix}

\item{ancestor}{Ancestor node}

\item{depth}{Maximum depth}

\item{npc}{Number of PCs}

\item{smooth}{Whether to smooth curves}
}
\description{
This function is a helper function used to make a new split in a FunFor tree
}
