% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunFor.R
\name{mvTree}
\alias{mvTree}
\title{Fit an FunFor tree}
\usage{
mvTree(formula, data, npc = NULL, m_split, smooth)
}
\arguments{
\item{formula}{Formula of model fitted}

\item{data}{All the data}

\item{npc}{Number of PCs}

\item{m_split}{Optimal tree size}

\item{smooth}{Whether to smooth curves}
}
\description{
This function is used to fit an FunFor tree. It can be called independently or used in the main function FunFor().
}
\examples{
mvTree(Y ~ ., data, m_split = 20, smooth = TRUE)
}
