% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunFor.R
\name{predict.mvRF}
\alias{predict.mvRF}
\title{Make predictions based on a FunFor fit}
\usage{
\method{predict}{mvRF}(fit, newdata)
}
\arguments{
\item{fit}{Saved FunFor object}

\item{newdata}{A new X data matrix}
}
\description{
This function is to make predictions based on a FunFor fit.
}
\examples{
predict(funfor_fit, newX)
}
