% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunFor.R
\name{predict.mvTree}
\alias{predict.mvTree}
\title{Make predictions based on a fitted tree}
\usage{
\method{predict}{mvTree}(fit, newdata)
}
\arguments{
\item{fit}{Saved tree object}

\item{newdata}{A new X data matrix}
}
\description{
This function is to make predictions based on a fitted tree.
}
\examples{
predict(tree_fit, newX)
}
