% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunFor.R
\name{search_tree}
\alias{search_tree}
\title{A helper function for FunFor prediction}
\usage{
search_tree(MVT, X, start.node)
}
\arguments{
\item{MVT}{A helper matrix}

\item{X}{X matrix}

\item{start.node}{Starting node}
}
\description{
This function is called by predict.mvTree() when making an prediction.
}
